简化依赖PIR
