+++ java/ppc-pir-services/sdk/src/main/java/com/wedpr/pir/sdk/PirClient.java
@@ -1,8 +1,6 @@
 package com.wedpr.pir.sdk;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.sun.org.slf4j.internal.Logger;
-import com.sun.org.slf4j.internal.LoggerFactory;
 import com.wedpr.pir.sdk.entity.body.SimpleEntity;
 import com.wedpr.pir.sdk.entity.param.PirJobParam;
 import com.wedpr.pir.sdk.entity.request.ServerJobRequest;
@@ -19,7 +17,6 @@ import java.util.Objects;
  * @date 2024/8/19
  */
 public class PirClient {
-    private static final Logger logger = LoggerFactory.getLogger(PirClient.class);
     private final PirJobService pirJobService;
 
     public PirClient(PirJobParam pirJobParam) throws WedprException {
@@ -33,21 +30,18 @@ public class PirClient {
             PirJobParam pirJobParam = pirJobService.getPirJobParam();
             ServerJobRequest serverJobRequest =
                     generatePirJobRequestFromJobAndOTParam(pirJobParam, otParamResponse);
-            logger.debug("executePirJob start: ", serverJobRequest);
 
             String pirUrl = pirJobParam.getGatewayUrl();
             ObjectMapper objectMapper = new ObjectMapper();
             String body = objectMapper.writeValueAsString(serverJobRequest);
             SimpleEntity otResult =
                     GatewayClient.sendPostRequestWithRetry(pirUrl, body, SimpleEntity.class);
-            logger.debug("executePirJob success return: ", otResult);
             if (Objects.isNull(otResult)
                     || !otResult.getCode().equals(WedprStatusEnum.SUCCESS.getCode())) {
                 throw new WedprException(WedprStatusEnum.TASK_EXECUTE_ERROR);
             }
             return pirJobService.requesterOtRecover(otParamResponse, pirJobParam, otResult);
         } catch (Exception e) {
-            logger.error("executePirJob cause exception: ", e.getMessage());
             throw new WedprException(WedprStatusEnum.CALL_PSI_ERROR);
         }
     }
