+++ java/ppc-pir-services/sdk/src/main/java/com/wedpr/pir/sdk/helper/CryptoOperatorHelper.java
@@ -1,7 +1,15 @@
 package com.wedpr.pir.sdk.helper;
 
+import com.wedpr.pir.sdk.exception.WedprException;
+import com.wedpr.pir.sdk.exception.WedprStatusEnum;
+
 import java.math.BigInteger;
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
 import java.security.SecureRandom;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.UUID;
 
 /**
  * @author zachma
@@ -47,5 +55,44 @@ public class CryptoOperatorHelper {
     /** a*b mod N */
     public static BigInteger OTMul(BigInteger a, BigInteger b) {
         return a.multiply(b).mod(DEFAULT_N);
+    }
+
+    /**
+     * 批量获取searchID的Hash值
+     * */
+    public static List<String> getIdHashVec(int obfuscationOrder, Integer idIndex, String searchId)
+            throws WedprException {
+
+        List<String> clientDataBodyList = new ArrayList<>();
+        String searchIdTemp;
+        for (int i = 0; i < obfuscationOrder + 1; i++) {
+            if (idIndex == i) {
+                searchIdTemp = searchId;
+            } else {
+                String uuid = UUID.randomUUID().toString();
+                searchIdTemp = makeHash(uuid.getBytes());
+            }
+            clientDataBodyList.add(searchIdTemp);
+        }
+        return clientDataBodyList;
+    }
+
+    private static String makeHash(byte[] data) throws WedprException {
+        try {
+            MessageDigest digest = MessageDigest.getInstance("SHA-256");
+            byte[] hashBytes = digest.digest(data);
+
+            StringBuilder hexString = new StringBuilder();
+            for (byte b : hashBytes) {
+                String hex = Integer.toHexString(0xFF & b);
+                if (hex.length() == 1) {
+                    hexString.append('0');
+                }
+                hexString.append(hex);
+            }
+            return hexString.toString();
+        } catch (NoSuchAlgorithmException e) {
+            throw new WedprException(WedprStatusEnum.WRONG_HASH_ERROR);
+        }
     }
 }
